Move CommandException to top level exception package
