+++ src/main/java/seedu/address/logic/commands/AddCommand.java
@@ -7,7 +7,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.exception.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
