+++ src/main/java/seedu/address/logic/commands/Command.java
@@ -1,6 +1,6 @@
 package seedu.address.logic.commands;
 
-import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.exception.CommandException;
 import seedu.address.model.Model;
 
 /**
