+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.exception.CommandException;
 import seedu.address.exception.NotFoundException;
-import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
